5lecture/requirements.txt
fastapi
uvicorn
google-cloud-storage
python-multipart
google-resumable-media